move the DsMap and CftMap into CCutils
