+++ mrmustard/lab_dev/transformations/transformations.py
@@ -354,64 +354,3 @@ class Attenuator(Channel):
         n_modes = len(self.modes)
         eta = list(reshape_params(n_modes, eta=self.transmissivity.value))[0]
         return Bargmann(*triples.attenuator_Abc(eta))
-
-
-class D_sMap(Channel):
-    r"""The s-parametrized Dgate as a channel.
-    This will be used as an internal Channel for representation transformation.
-
-    Args:
-        num_modes: number of modes of this channel.
-        s: s-parameter of this channel.
-
-    .. details::
-
-        Its ``(A,b,c)`` triple is given by 
-
-        .. math::
-            A &= \\
-            b &= \\
-            c &= 1\:.
-    """
-
-    def __init__(
-        self,
-        modes: Sequence[int],
-        s: float,
-    ):
-        super().__init__(modes=modes, name="DsMap")
-        self.s = s
-
-    @property
-    def representation(self) -> Bargmann:
-        n_modes = len(self.modes)
-        return Bargmann(*triples.displacement_map_s_parametrized_Abc(self.s, n_modes))
-
-
-class CFMap(Channel):
-    r"""The Complex Fourier Transformation as a channel.
-    This will be used as an internal Channel for representation transformation.
-
-    Args:
-        num_modes: number of modes of this channel.
-
-    .. details::
-
-        Its ``(A,b,c)`` triple is given by 
-
-        .. math::
-            A &= \\
-            b &= \\
-            c &= 1\:.
-    """
-
-    def __init__(
-        self,
-        modes: Sequence[int],
-    ):
-        super().__init__(modes=modes, name="CFMap")
-
-    @property
-    def representation(self) -> Bargmann:
-        n_modes = len(self.modes)
-        return Bargmann(*triples.complex_fourier_transform_Abc(n_modes))
